+++ weblate/checks/consistency.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 
+from collections import defaultdict
 from functools import reduce
 from typing import TYPE_CHECKING, Literal
 
@@ -15,6 +16,8 @@ from weblate.checks.base import BatchCheckMixin, TargetCheck
 from weblate.utils.state import STATE_TRANSLATED
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
+
     from weblate.trans.models import Component, Unit
     from weblate.trans.models.unit import UnitQuerySet
 
@@ -110,7 +113,7 @@ class ConsistencyCheck(TargetCheck, BatchCheckMixin):
         """Target strings are checked in check_target_unit."""
         return False
 
-    def check_component(self, component: Component):
+    def check_component(self, component: Component) -> Iterable[Unit]:
         from weblate.trans.models import Unit
 
         units = Unit.objects.filter(
@@ -170,8 +173,14 @@ class ReusedCheck(TargetCheck, BatchCheckMixin):
         from weblate.trans.models import Unit
 
         if target is None:
-            return unit.same_target_units
-        return Unit.objects.same_target(unit, target)
+            result = unit.same_target_units
+        else:
+            result = Unit.objects.same_target(unit, target)
+
+        if not unit.translation.language.is_case_sensitive():
+            result = result.exclude(source__lower=unit.source.lower())
+
+        return result
 
     def should_skip(self, unit: Unit):
         if unit.translation.plural.number <= 1 or not any(unit.get_target_plurals()):
@@ -203,7 +212,7 @@ class ReusedCheck(TargetCheck, BatchCheckMixin):
         """Target strings are checked in check_target_unit."""
         return False
 
-    def check_component(self, component: Component):
+    def check_component(self, component: Component) -> Iterable[Unit]:
         from weblate.trans.models import Unit
 
         units = Unit.objects.filter(
@@ -229,9 +238,9 @@ class ReusedCheck(TargetCheck, BatchCheckMixin):
         )
 
         if not matches:
-            return []
+            return
 
-        return (
+        result = (
             units.filter(
                 reduce(
                     lambda x, y: x
@@ -248,6 +257,22 @@ class ReusedCheck(TargetCheck, BatchCheckMixin):
             .prefetch()
             .prefetch_bulk()
         )
+
+        # Filter out case differing source for case insensitive languages
+        found: dict[tuple(str, str), set[str]] = defaultdict(set)
+        remaining: list[tuple[str, Unit]] = []
+        for unit in result:
+            if not unit.translation.language.is_case_sensitive():
+                key = (unit.translation.language.code, unit.target)
+                lower_source = unit.source.lower()
+                found[key].add(lower_source)
+                remaining.append((key, unit))
+            else:
+                yield unit
+
+        for key, unit in remaining:
+            if len(found[key]) > 1:
+                yield unit
 
 
 class TranslatedCheck(TargetCheck, BatchCheckMixin):
@@ -325,7 +350,7 @@ class TranslatedCheck(TargetCheck, BatchCheckMixin):
             return None
         return [(".*", target, "u")]
 
-    def check_component(self, component: Component):
+    def check_component(self, component: Component) -> Iterable[Unit]:
         from weblate.trans.models import Change, Unit
 
         units = (